Move OrchestrationException to api package
